+++ packages/kit-bg/src/services/ServiceScanQRCode/index.test.ts
@@ -1,6 +1,7 @@
+import { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
+
 import { parseQRCode as parse } from './utils/parseQRCode';
 import { PARSE_HANDLER_NAMES } from './utils/parseQRCode/handlers';
-import { EQRCodeHandlerType } from './utils/parseQRCode/type';
 
 // yarn jest packages/kit-bg/src/services/ServiceScanQRCode/index.test.ts
 describe('useParseQRCode', () => {
