+++ packages/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/handlers/sendProtection.ts
@@ -3,7 +3,7 @@ import {
   WEB_APP_URL_DEV,
 } from '@onekeyhq/shared/src/config/appConfig';
 
-import { EQRCodeHandlerType } from '../type';
+import { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
 
 import type { IBaseValue, IQRCodeHandler } from '../type';
