+++ packages/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/handlers/solana.ts
@@ -1,6 +1,6 @@
 import { getNetworkIdsMap } from '@onekeyhq/shared/src/config/networkIds';
 
-import { EQRCodeHandlerType } from '../type';
+import { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
 
 import type { IQRCodeHandler, ISolanaValue } from '../type';
