+++ packages/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/handlers/url.ts
@@ -1,6 +1,6 @@
 import { parseUrl } from '@onekeyhq/shared/src/utils/uriUtils';
 
-import { EQRCodeHandlerType } from '../type';
+import { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
 
 import type { IQRCodeHandler, IUrlValue } from '../type';
