+++ packages/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/index.ts
@@ -1,9 +1,9 @@
 import { defaultLogger } from '@onekeyhq/shared/src/logger/logger';
+import { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
 
 import { PARSE_HANDLERS, PARSE_HANDLER_NAMES } from './handlers';
 import * as deeplinkHandler from './handlers/deeplink';
 import * as urlHandler from './handlers/url';
-import { EQRCodeHandlerType } from './type';
 
 import type {
   IBaseValue,
