+++ packages/shared/types/qrCode.ts
@@ -1,0 +1,15 @@
+export enum EQRCodeHandlerType {
+  UNKNOWN = 'UNKNOWN',
+  BITCOIN = 'BITCOIN',
+  ETHEREUM = 'ETHEREUM',
+  SOLANA = 'SOLANA',
+  LIGHTNING_NETWORK = 'LIGHTNING_NETWORK',
+  URL = 'URL',
+  WALLET_CONNECT = 'WALLET_CONNECT',
+  MIGRATE = 'MIGRATE',
+  ANIMATION_CODE = 'ANIMATION_CODE',
+  DEEPLINK = 'DEEPLINK',
+  URL_ACCOUNT = 'URL_ACCOUNT',
+  MARKET_DETAIL = 'MARKET_DETAIL',
+  SEND_PROTECTION = 'SEND_PROTECTION',
+}
